==> JSW/01.ex_year.ipynb <==
병합시킨 파일 불러오기 

import numpy as np 
import pandas as pd

merged_df = pd.read_csv('merged_df.csv')

# merged_df
# merged_df.info()

# merged_df.columns

# 행정동 컬럼 확인 
# merged_df['행정동'].unique()

# 동 갯수 확인 
# merged_dong = merged_df['행정동'].nunique()

# merged_dong # 행정동 426개, 4개가 비었음

# NaN값 확인 (0.0)
# zero_count = (merged_df['지하철_승객_수'] == 0.0).sum()

# zero_count 

연도별 데이터 확인 

# # 연도별 데이터 확인
# merged_df['연도'] = merged_df['기준_날짜'].astype(str).str[:4]
# merged_df[['기준_날짜', '연도']]

연도 컬럼 추출

# # 연도별 (2022)
# merged_df_2022 = merged_df[merged_df['연도'] == '2022']
# merged_df_2022

# # 연도별 (2023)
# merged_df_2023 = merged_df[merged_df['연도'] == '2023']
# merged_df_2023

# # 연도별 (2024)
# merged_df_2024 = merged_df[merged_df['연도'] == '2024']
# merged_df_2024

# print(merged_df['연도'].unique())

# # 연도 리스트 
# years = ['2024','2023','2022']

# year = {year:merged_df[merged_df['연도'] == year] for year in years}

# year['2024']
# year['2022']
# year['2023']

# # 각각 년도 (2024, 2023, 2022, 2021) 데이터프레임 저장
# for year in years:
#     year_24 = merged_df[merged_df['연도'] == year]
#     filename = f'data_{year}.csv'
#     year_24.to_csv(filename, index=False, encoding='utf-8')
#     (f"{filename} 저장 완료!")
