==> JSW/1.zone+subway.ipynb <==
데이터 불러오기
대충 전처리

import numpy as np 
import pandas as pd

file = "행정동_코드.xlsx"

z = pd.read_excel('file/행정동_코드.xlsx', sheet_name='sheet1')

z.head()

z.columns

# 필요없는 컬럼 삭제 
columns_d = ["시도","행정동\n(행정기관명)","법정동","행정기관 생성일","법정동코드",
             "법정동 관할구역\n분할여부","행정동 영문명칭","비고","Unnamed: 12","내"]
z.drop(columns=columns_d, inplace=True)

z.columns

z.rename(columns={'행정구역분류' : '행정동_ID','행정구역명' : '행정동'}, inplace=True)

z.columns

z.head()

# 행정동 확인
z['행정동'].unique()

# z.to_csv("z.csv", index=False, encoding="utf-8")

subway_df = pd.read_csv('file/지하철.csv', encoding="CP949")
subway_df

# 필요 컬럼 남기기
subway_df.columns
subway_df = subway_df[['기준_날짜', '행정동_ID', '지하철_승객_수']]
subway_df

# NaN값 0으로 채우기
merged_df_cleaned = subway_df.fillna(0)
merged_df_cleaned

# Csv파일로 저장하기
# merged_df_cleaned.to_csv('sb.csv', index=False, encoding='utf-8')
# print('CSV파일 저장 완료!')
