==> JSW/3.year_22_23_24.ipynb <==
각각 22년, 23년, 24년 데이터 가공

import numpy as np 
import pandas as pd

year22 = pd.read_csv('data_2022.csv')
year23 = pd.read_csv('data_2023.csv')
year24 = pd.read_csv('data_2024.csv')

year22
year23
year24

# '기준_날짜'를 문자열로 변환하기
year22['기준_날짜'] = year22['기준_날짜'].astype(str) 
year22['기준_날짜'] = pd.to_datetime(year22['기준_날짜'],format='%Y%m%d').dt.strftime('%Y%m%d')
year22['기준_날짜']

# '기준_날짜'를 문자열로 변환하기
year23['기준_날짜'] = year23['기준_날짜'].astype(str) 
year23['기준_날짜'] = pd.to_datetime(year23['기준_날짜'],format='%Y%m%d').dt.strftime('%Y%m%d')
year23['기준_날짜']

# '기준_날짜'를 문자열로 변환하기
year24['기준_날짜'] = year24['기준_날짜'].astype(str) 
year24['기준_날짜'] = pd.to_datetime(year24['기준_날짜'],format='%Y%m%d').dt.strftime('%Y%m%d')
year24['기준_날짜']

날짜 변환

# YYYY-MM-DD 형식으로 변환
year22['기준_날짜'] = year22['기준_날짜'].astype(str)
year22['기준_날짜'] = pd.to_datetime(year22['기준_날짜'], format='%Y%m%d').dt.strftime('%Y년%m월%d일')
print(year22['기준_날짜'].head())# 결과 확인


# YYYY-MM-DD 형식으로 변환
year23['기준_날짜'] = year23['기준_날짜'].astype(str)
year23['기준_날짜'] = pd.to_datetime(year23['기준_날짜'], format='%Y%m%d').dt.strftime('%Y년%m월%d일')
print(year23['기준_날짜'].head())


# YYYY-MM-DD 형식으로 변환
year24['기준_날짜'] = year24['기준_날짜'].astype(str)
year24['기준_날짜'] = pd.to_datetime(year24['기준_날짜'], format='%Y%m%d').dt.strftime('%Y년%m월%d일')
print(year24['기준_날짜'].head())

year22
year23
year24

year22.to_csv('year22.csv', index=False, encoding='utf-8')
year23.to_csv('year23.csv', index=False, encoding='utf-8')
year24.to_csv('year24.csv', index=False, encoding='utf-8')
