==> JSW/04.ex_month_out.ipynb <==
년도별 가공한 파일 3개 불러오기

import numpy as np 
import pandas as pd

# 파일 3개 불러오기
year22 = pd.read_csv('..//makefile2-1/year22.csv')
year23 = pd.read_csv('../makefile2-1/year23.csv')
year24 = pd.read_csv('../makefile2-1/year24.csv')

# year22['기준_날짜'].dtype
# year23['기준_날짜'].dtype
# year24['기준_날짜'].dtype

2022년

# # 2022년 
# year22 = pd.DataFrame(year22)

# # '기준_날짜'에서 월추출
# year22['월'] = year22['기준_날짜'].str[5:7] # '2021년11월16일에서 '11' 추출

# # 각 월별로 필터링
# month22 = {month: year22[year22['월'] == month] for month in ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12']}

# year22

# year22['월'].unique()

# # group by
# year22_sum = year22.groupby(['연도', '월', '행정동','시군구'])['지하철_승객_수'].sum().reset_index()
# year22_sum

년도별 분기 생성

# # 분기별 생성
# def get_quarter(month):
#     if month in ['01', '02', '03']:
#         return '1-3월'
#     elif month in ['04','05','06']:
#         return '4-6월'
#     elif month in ['07','08','09']:
#         return '7-9월'
#     else:
#         return '10-12월'

# year22['분기'] = year22['월'].apply(get_quarter)

# # 분기별로 groupby, 지하철 승객 수 합계 계산 
# year22_sp = year22.groupby(['연도','분기','행정동','시군구'])['지하철_승객_수'].sum().reset_index()
# year22_sp

# # '지하철_승객_수'가 0.0 확인
# zero_count = year22[year22['지하철_승객_수'] == 0.0].shape[0]
# (f"지하철 승객 수가 0.0인 행의 개수: {zero_count}")

# year22_sp.to_csv('y22_quarter_sum.csv', index=False)

2023년


# # 2023년
# year23 = pd.DataFrame(year23)

# # '기준_날짜'에서 월추출
# year23['월'] = year23['기준_날짜'].str[5:7] # '2021년11월16일에서 '11' 추출

# # 각 월별로 필터링
# month23 = {month: year23[year23['월'] == month] for month in ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12']}

# year23

# year23['월'].unique()

# year23_sum = year23.groupby(['연도', '월', '행정동','시군구'])['지하철_승객_수'].sum().reset_index()
# year23_sum

# # 분기별 생성
# def get_quarter(month):
#     if month in ['01', '02', '03']:
#         return '1-3월'
#     elif month in ['04','05','06']:
#         return '4-6월'
#     elif month in ['07','08','09']:
#         return '7-9월'
#     else:
#         return '10-12월'

# year23['분기'] = year23['월'].apply(get_quarter)

# # 분기별로 groupby, 지하철 승객 수 합계 계산 
# year23_sp = year23.groupby(['연도','분기','행정동','시군구'])['지하철_승객_수'].sum().reset_index()
# year23_sp

# year23_sp.to_csv('y23_quarter_sum.csv', index=False)

2024년


# # 2024년
# year24 = pd.DataFrame(year24)

# # '기준_날짜'에서 월추출
# year24['월'] = year24['기준_날짜'].str[5:7] # '2021년11월16일에서 '11' 추출

# # 각 월별로 필터링
# month24 = {month: year24[year24['월'] == month] for month in ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12']}

# year24

# year24['월'].unique()

# year24_sum = year24.groupby(['연도', '월', '행정동','시군구'])['지하철_승객_수'].sum().reset_index()
# year24_sum

# # 분기별 생성
# def get_quarter(month):
#     if month in ['01', '02', '03']:
#         return '1-3월'
#     elif month in ['04','05','06']:
#         return '4-6월'
#     elif month in ['07','08','09']:
#         return '7-9월'
#     else:
#         return '10-11월'

# year24_sum['분기'] = year24_sum['월'].apply(get_quarter)

# # 분기별로 groupby, 지하철 승객 수 합계 계산 
# year24_sp = year24_sum.groupby(['연도','분기','행정동','시군구'])['지하철_승객_수'].sum().reset_index()
# year24_sp

# 10-11월 확인
year24_o = year24_sp[year24_sp['분기'] == '10-11월']
year24_o

year22_sp['행정동'].unique()
# year23_sp['행정동'].unique()
# year24_sp['행정동'].unique()

year24_sp.to_csv('y24_quarter_sum.csv', index=False)
