==> JSW/06.ex_quarter_drop.ipynb <==
년도별로 승객수가 있는 데이터와 없는 데이터로 분할

import numpy 
import pandas as pd

# 파일불러오기
is22 = pd.read_csv('../makefile2-2/year_2022_is_dong.csv')
is23 = pd.read_csv('../makefile2-2/year_2023_is_dong.csv')
is24 = pd.read_csv('../makefile2-2/year_2024_is_dong.csv')

zero22 = '../makefile2-2/year_2022_zero_dong.csv'
zero23 = '../makefile2-2/year_2023_zero_dong.csv'
zero24 = '../makefile2-2/year_2024_zero_dong.csv'

no22 = pd.read_csv(zero22)
no23 = pd.read_csv(zero23)
no24 = pd.read_csv(zero24)

# no22

# print(no24.isna().sum())

# # 년도별 데이터에 행정동 갯수,목록 확인

# is22['행정동'].unique()
# is22['행정동'].nunique()

# is23['행정동'].unique()
# is23['행정동'].nunique()

# is24['행정동'].unique()
# is24['행정동'].nunique()

# # 22 : 219
# # 23 : 219
# # 24 : 221

행정동 존재 여부 확인

# is22['행정동'].unique()
# print("2022년 데이터에 '항동' 존재 여부:", "항동" in is22)

# is23['행정동'].unique()
# print("2022년 데이터에 '항동' 존재 여부:", "항동" in is23)

# # 2022-2023, 2023-2024, 2022-2024
# # 년도별 공통된 행정동목록과 차이가 있는 행정동목록 확인

# # 문자열로 변환
# dong22 = set(is22['행정동'].astype(str).unique())
# dong23 = set(is23['행정동'].astype(str).unique())
# dong24 = set(is24['행정동'].astype(str).unique())

# # 2022-2023

# # 공통된 행정동 (교집합) 
# samedong_22_23 = dong22 & dong23 

# # 다른 행정동 (공통된 행정동 제거)
# diffdong_22_23 = (dong22 | dong23) - samedong_22_23

# print(f"공통된 행정동({len(samedong_22_23)}개):")
# print(samedong_22_23)

# print(f"\n다른 행정동 ({len(diffdong_22_23)}개):")
# print(diffdong_22_23)

# # 공통된 행정동(219)
# # 다른 행정동(X)

# # 2023-2024

# # 공통된 행정동 (교집합) 
# samedong_23_24 = dong23 & dong24

# # 다른 행정동 (공통된 행정동 제거)
# diffdong_23_24 = (dong23 | dong24) - samedong_23_24

# print(f"공통된 행정동({len(samedong_22_23)}개):")
# print(samedong_23_24)

# print(f"\n다른 행정동 ({len(diffdong_23_24)}개):")
# print(diffdong_23_24)

# # 공통된 행정동(219)
# # 다른 행정동(2개) : {'오류2동', '암사3동'}

# 2022 2024


# # 공통된 행정동 (교집합) 
# samedong_22_24 = dong22 & dong24

# # 다른 행정동 (공통된 행정동 제거)
# diffdong_22_24 = (dong22 | dong24) - samedong_22_24

# print(f"공통된 행정동({len(samedong_22_23)}개):")
# print(samedong_22_24)

# print(f"\n다른 행정동 ({len(diffdong_22_24)}개):")
# print(diffdong_22_24)

# # 공통된 행정동(219)
# # 다른 행정동(2개) : {'오류2동', '암사3동','항동'}

# # 결론 : 2024년에 3개가 다름

# print(is24.dtypes)

# # 삭제할 컬럼명
# c_drop = '분기'

# if c_drop in no22.columns:
#     no22 = no22.drop(columns=[c_drop])

# if c_drop in no23.columns:
#     no23 = no23.drop(columns=[c_drop])

# if c_drop in no24.columns:
#     no24 = no24.drop(columns=[c_drop])

# # 파일로 다시 저장 
# no22.to_csv('year_2022_zero_dong.csv', index=False)
# no23.to_csv('year_2023_zero_dong.csv', index=False)
# no24.to_csv('year_2024_zero_dong.csv', index=False)

# print(f"컬럼 '{c_drop} 가 삭제되었습니다.")

# is22 
# no22

groupby 해서 총_승객수 평균값 구하기

# is22.groupby(['연도', '분기', '행정동', '시군구'])['총_승객수'].mean()
# is23.groupby(['연도', '분기', '행정동', '시군구'])['총_승객수'].mean()
# is24.groupby(['연도', '분기', '행정동', '시군구'])['총_승객수'].mean()

# # is_years = [is22, is23, is24]
# # results = [df.groupby(['연도','분기','행정동'])['총_승객수'].mean().reset_index() for df in is_years]

# # is22 데이터프레임 그룹화 후 CSV로 저장
# is22_grouped = is22.groupby(['연도', '분기', '행정동', '시군구'])['총_승객수'].mean().reset_index()
# is22_grouped.to_csv('is22_grouped.csv', index=False)

# # is23 데이터프레임 그룹화 후 CSV로 저장
# is23_grouped = is23.groupby(['연도', '분기', '행정동', '시군구'])['총_승객수'].mean().reset_index()
# is23_grouped.to_csv('is23_grouped.csv', index=False)

# # is24 데이터프레임 그룹화 후 CSV로 저장
# is24_grouped = is24.groupby(['연도', '분기', '행정동', '시군구'])['총_승객수'].mean().reset_index()
# is24_grouped.to_csv('is24_grouped.csv', index=False)

# # 데이터프레임 합치기 
# year22_df_all = pd.concat([is22_grouped, no22],ignore_index=True)
# year23_df_all = pd.concat([is23_grouped, no23],ignore_index=True)
# year24_df_all = pd.concat([is24_grouped, no24],ignore_index=True)

# year22_df_all.head()
# year23_df_all.head()
# year24_df_all.head()

# 데이터프레임 저장 
year22_df_all.to_csv('year22_df_all.csv', index=False)
year23_df_all.to_csv('year23_df_all.csv', index=False)
year24_df_all.to_csv('year24_df_all.csv', index=False)
