==> JSW/7.quarter_top.ipynb <==
import numpy as np 
import pandas as pd 

year22 = pd.read_csv('file/year22_df_all.csv')
year23 = pd.read_csv('file/year23_df_all.csv')
year24 = pd.read_csv('file/year24_df_all.csv')

year22, year23, year24

year22['행정동'].unique()
year23['행정동'].unique()
year24['행정동'].unique()

print(year24.isna().sum())

# 항동, 상일동에 데이터 있는지 확인

hang_dong = year24[year24['행정동'].str.contains('항동')]

print("행정동 목록에서 '항동'이 포함된 동:")
print(hang_dong)

# 결과: 둘다 행정동 데이터 없음

# 총_승객수 값이 0.0 제외 
year22_df = year22[year22['총_승객수'] != 0.0]
year22_df

year23_df = year23[year23['총_승객수'] != 0.0]
year23_df

year24_df = year24[year24['총_승객수'] != 0.0]
year24_df

# 다른구 같은 동 확인
def same_dong(df, dong_name, gu_name):
    same_dong = df[df['행정동'] == dong_name] 
    other_gus = same_dong[same_dong['시군구'] !=gu_name] 
    unique_gus = other_gus[['시군구','행정동']].drop_duplicates().reset_index(drop=True)
    return unique_gus 

dong_name = "신사동"
gu_name = "강남구"

same_dong_in_other_gus = same_dong(year22_df, dong_name, gu_name)
print(f"'{dong_name}'이(가) '{gu_name}' 외에 속한 구 목록:")
print(same_dong_in_other_gus)

# shinsadong = year22_df[year22_df['행정동'] == '신사동'] 

# print("신사동 목록:")
# print(shinsadong)

shinsadong = year22_df[year22_df['시군구'] == '강남구'] 

print("강남구 목록:")
print(shinsadong)

# --분기별로 어느 행정동이 상위에 드는지 확인 (2022년)

# 그룹화 : 연도, 분기, 행정동 별로 총_승객수 합산
dong23 = year22_df.groupby(['연도', '분기', '행정동'])['총_승객수'].sum().reset_index()

# 2022년, 분기별로 총_승객수가 가장 많은 행정동 찾기
top_dong23 = dong23.loc[dong23.groupby(['연도','분기'])['총_승객수'].idxmax()]
top_dong23 = top_dong23.reset_index(drop=True)

print(top_dong23)


# --분기별로 어느 행정동이 상위에 드는지 확인 (2023년)

# 그룹화 : 연도, 분기, 행정동 별로 총_승객수 합산
dong23 = year23_df.groupby(['연도', '분기', '행정동'])['총_승객수'].sum().reset_index()

# 2022년, 분기별로 총_승객수가 가장 많은 행정동 찾기
top_dong23 = dong23.loc[dong23.groupby(['연도','분기'])['총_승객수'].idxmax()]
top_dong23 = top_dong23.reset_index(drop=True)

print(top_dong23)


# --분기별로 어느 행정동이 상위에 드는지 확인 (2024년)

# 그룹화 : 연도, 분기, 행정동 별로 총_승객수 합산
dong24 = year24_df.groupby(['연도', '분기', '행정동'])['총_승객수'].sum().reset_index()

# 2022년, 분기별로 총_승객수가 가장 많은 행정동 찾기
top_dong24 = dong24.loc[dong24.groupby(['연도','분기'])['총_승객수'].idxmax()]
top_dong24 = top_dong24.reset_index(drop=True)

print(top_dong24)

2022년 분기별 행정동 상위 순위

# --2022년 1분기 행정동 승객_수 top 5--

dong22_q1 = year22_df[(year22_df['연도'] == 2022) & (year22_df['분기'] == '1분기')]
top5_dong22_q1 = dong22_q1.nlargest(11, '총_승객수') 
top5_dong22_q1 = top5_dong22_q1.reset_index(drop=True)
print(top5_dong22_q1)

# --2022년 2분기 행정동 승객_수 top 5--
dong22_q2 = year22_df[(year22_df['연도'] == 2022) & (year22_df['분기'] == '2분기')]
top5_dong22_q2 = dong22_q2.nlargest(11, '총_승객수')
top5_dong22_q2 = top5_dong22_q2.reset_index(drop=True) 
print(top5_dong22_q2)

# --2022 3분기 행정동 승객_수 top 5--
dong22_q3 = year22_df[(year22_df['연도'] == 2022) & (year22_df['분기'] == '3분기')]
top5_dong22_q3 = dong22_q3.nlargest(11, '총_승객수')
top5_dong22_q3 = top5_dong22_q3.reset_index(drop=True)
print(top5_dong22_q3)

# --2022 4분기 행정동 승객_수 top 5--
dong22_q4 = year22_df[(year22_df['연도'] == 2022) & (year22_df['분기'] == '4분기')]
top5_dong22_q4 = dong22_q4.nlargest(11, '총_승객수')
top5_dong22_q4 = top5_dong22_q4.reset_index(drop=True)
print(top5_dong22_q4)

2022년 분기별 행정동 하위 순위

# 1분기
dong22_q1 = year22_df[(year22_df['연도'] == 2022) & (year22_df['분기'] == '1분기')]
bottom_dong22_q1 = dong22_q1.nsmallest(11, '총_승객수')
bottom_dong22_q1 = bottom_dong22_q1.reset_index(drop=True)
print(bottom_dong22_q1)

# 2분기
dong22_q2 = year22_df[(year22_df['연도'] == 2022) & (year22_df['분기'] == '2분기')]
bottom_dong22_q2 = dong22_q2.nsmallest(11, '총_승객수')
bottom_dong22_q2 = bottom_dong22_q2.reset_index(drop=True)
print(bottom_dong22_q2)

# 3분기
dong22_q3 = year22_df[(year22_df['연도'] == 2022) & (year22_df['분기'] == '3분기')]
bottom_dong22_q3 = dong22_q3.nsmallest(11, '총_승객수')
bottom_dong22_q3 = bottom_dong22_q3.reset_index(drop=True)
print(bottom_dong22_q3)

# 4분기
dong22_q3 = year22_df[(year22_df['연도'] == 2022) & (year22_df['분기'] == '3분기')]
bottom_dong22_q3 = dong22_q3.nsmallest(11, '총_승객수')
bottom_dong22_q3 = bottom_dong22_q3.reset_index(drop=True)
print(bottom_dong22_q3)

2023년 분기별 행정동 상위 순위


# 2023 1분기 행정동 승객_수 top 5
dong23_q1 = year23_df[(year23_df['연도'] == 2023) & (year23_df['분기'] == '1분기')]
top5_dong23_q1 = dong23_q1.nlargest(11, '총_승객수')
top5_dong23_q1 = top5_dong23_q1.reset_index(drop=True)
print(top5_dong23_q1)

# 2023 2분기 행정동 승객_수 top 5
dong23_q2 = year23_df[(year23_df['연도'] == 2023) & (year23_df['분기'] == '2분기')]
top5_dong23_q2 = dong23_q2.nlargest(11, '총_승객수')
top5_dong23_q2 = top5_dong23_q2.reset_index(drop=True)
print(top5_dong23_q2)

# 2023 3분기 행정동 승객_수 top 5
dong23_q3 = year23_df[(year23_df['연도'] == 2023) & (year23_df['분기'] == '3분기')]
top5_dong23_q3 = dong23_q3.nlargest(11, '총_승객수')
top5_dong23_q3 = top5_dong23_q3.reset_index(drop=True)
print(top5_dong23_q3)

# 2023 4분기 행정동 승객_수 top 5
dong23_q4 = year23_df[(year23_df['연도'] == 2023) & (year23_df['분기'] == '4분기')]
top5_dong23_q4 = dong23_q4.nlargest(11, '총_승객수')
top5_dong23_q4 = top5_dong23_q4.reset_index(drop=True)
print(top5_dong23_q4)

2023년 분기별 행정동 하위 순위

# 2023년 1분기 하위 5개 행정동
dong23_q1 = year23_df[(year23_df['연도'] == 2023) & (year23_df['분기'] == '1분기')]
bottom5_dong23_q1 = dong23_q1.nsmallest(5, '총_승객수').reset_index(drop=True)
print("2023년 1분기 하위 5개 행정동:")
print(bottom5_dong23_q1)

# 2023년 2분기 하위 5개 행정동
dong23_q2 = year23_df[(year23_df['연도'] == 2023) & (year23_df['분기'] == '2분기')]
bottom5_dong23_q2 = dong23_q2.nsmallest(5, '총_승객수').reset_index(drop=True)
print("2023년 2분기 하위 5개 행정동:")
print(bottom5_dong23_q2)

# 2023년 3분기 하위 5개 행정동
dong23_q3 = year23_df[(year23_df['연도'] == 2023) & (year23_df['분기'] == '3분기')]
bottom5_dong23_q3 = dong23_q3.nsmallest(5, '총_승객수').reset_index(drop=True)
print("2023년 3분기 하위 5개 행정동:")
print(bottom5_dong23_q3)

# 2023년 4분기 하위 5개 행정동
dong23_q4 = year23_df[(year23_df['연도'] == 2023) & (year23_df['분기'] == '4분기')]
bottom5_dong23_q4 = dong23_q4.nsmallest(5, '총_승객수').reset_index(drop=True)
print("2023년 4분기 하위 5개 행정동:")
print(bottom5_dong23_q4)

2024년 분기별 행정동 상위 순위

# 2024 1분기 행정동 승객_수 top 5
dong24_q1 = year24_df[(year24_df['연도'] == 2024) & (year24_df['분기'] == '1분기')]
top5_dong24_q1 = dong24_q1.nlargest(11, '총_승객수')
top5_dong24_q1 = top5_dong24_q1.reset_index(drop=True)
print(top5_dong24_q1)

# 2024 2분기 행정동 승객_수 top 5
dong24_q2 = year24_df[(year24_df['연도'] == 2024) & (year24_df['분기'] == '2분기')]
top5_dong24_q2 = dong24_q2.nlargest(11, '총_승객수')
top5_dong24_q2 = top5_dong24_q2.reset_index(drop=True)
print(top5_dong24_q2)

# 2024 분기 행정동 승객_수 top 5
dong24_q3 = year24_df[(year24_df['연도'] == 2024) & (year24_df['분기'] == '3분기')]
top5_dong24_q3 = dong24_q3.nlargest(11, '총_승객수')
top5_dong24_q3 = top5_dong24_q3.reset_index(drop=True)
print(top5_dong24_q3)

2024년 분기별 행정동 하위 순위

# 2024년 1분기 하위 5개 행정동
dong24_q1 = year24_df[(year24_df['연도'] == 2024) & (year24_df['분기'] == '1분기')]
bottom5_dong24_q1 = dong24_q1.nsmallest(5, '총_승객수').reset_index(drop=True)
print("2024년 1분기 하위 5개 행정동:")
print(bottom5_dong24_q1)

# 2024년 2분기 하위 5개 행정동
dong24_q2 = year24_df[(year24_df['연도'] == 2024) & (year24_df['분기'] == '2분기')]
bottom5_dong24_q2 = dong24_q2.nsmallest(5, '총_승객수').reset_index(drop=True)
print("2024년 2분기 하위 5개 행정동:")
print(bottom5_dong24_q2)

# 2024년 3분기 하위 5개 행정동
dong24_q3 = year24_df[(year24_df['연도'] == 2024) & (year24_df['분기'] == '3분기')]
bottom5_dong24_q3 = dong24_q3.nsmallest(5, '총_승객수').reset_index(drop=True)
print("2024년 3분기 하위 5개 행정동:")
print(bottom5_dong24_q3)

# 2024년 4분기 하위 5개 행정동
dong24_q4 = year24_df[(year24_df['연도'] == 2024) & (year24_df['분기'] == '4분기')]
bottom5_dong24_q4 = dong24_q4.nsmallest(5, '총_승객수').reset_index(drop=True)
print("2024년 4분기 하위 5개 행정동:")
print(bottom5_dong24_q4)

# 2024년 3,4분기 삭제 
year24_zero_q34 = year24[year24['분기'] != '3분기']
year24_zero_q34 = year24[year24['분기'] != '4분기']
print("3분기 데이터 삭제:")
print("4분기 데이터 삭제:")
print(year24_zero_q34)

# year24_zero_q34['분기'].isna().sum()
year24_zero_q34.to_csv('yeay24_zero_q3.csv', index=False, encoding='utf-8')
print('삭제 완료')

year22.columns

# # 데이터 병합 
# 공통컬럼 = ['연도', '분기', '행정동', '시군구', '총_승객수']
# merged_22_23 = pd.merge(year22,year23, on=공통컬럼, how='outer')
# merged_year_all = pd.merge(merged_22_23, year24_zero_q3, on=공통컬럼, how='outer')
# print(merged_year_all)

# # csv파일로 저장 
# merged_year_all.to_csv('merged_year_all.csv', index=False)
# print("병합된 데이터프레임을 'merged_year_all_csv' 파일로 저장함")

년분기 컬럼 생성

year_22_all = pd.read_csv('file/year22_df_all.csv')
year_23_all = pd.read_csv('file/year23_df_all.csv')
year_24_all = pd.read_csv('file/year24_df_all.csv')

def clean_quarter_column(df):
    if'분기' in df.columns:
        df['분기'] = df['분기'].str.extract('(\d+)')
    return df 

year_22_all = clean_quarter_column(year_22_all)
year_23_all = clean_quarter_column(year_23_all)
year_24_all = clean_quarter_column(year_24_all)

년분기생성, 연도 컬럼 삭제

year_22_all
year_23_all
year_24_all

def create_year_quarter_column(df):
    df['년분기'] = df['연도'].astype(str) + df['분기'].astype(str) 
    return df

year_22_all = create_year_quarter_column(year_22_all)
year_23_all = create_year_quarter_column(year_23_all)
year_24_all = create_year_quarter_column(year_24_all)

각각 년도별 데이터 병합

# 연도, 분기 빼기 
# year_22_all = year_22_all.drop(columns=['연도','분기'])
# year_23_all = year_23_all.drop(columns=['연도','분기'])
# year_24_all = year_24_all.drop(columns=['연도','분기'])

# 데이터 병합
merged_final = pd.concat([year_22_all, year_23_all, year_24_all], axis=0, ignore_index=True)
merged_final

# csv파일로 저장
merged_final.to_csv('file/merged_final_all.csv', index=False)
print("데이터 병합 완료")
