==> JSW/08.ex_dong_subway.ipynb <==
import numpy as np 
import pandas as pd 

dong_sw = pd.read_csv('dong_sw.csv')

dong_sw

# 지하철_승객_수 컬럼명 바꾸기 

# 컬럼 이름을 변경하려면 데이터프레임에서 직접 rename을 사용해야 합니다.
dong_sw.rename(columns={'지하철_승객_수': '총_승객수'}, inplace=True)

# 결과 확인
print(dong_sw)

dong_sw

year22 = dong_sw[dong_sw['연도'] == 2022]
year23 = dong_sw[dong_sw['연도'] == 2023]
year24 = dong_sw[dong_sw['연도'] == 2024]

year22
year23
year24

no22 = year22[year22['총_승객수'] == 0.00]
is22 = year22[year22['총_승객수'] !=0.00]

no23 = year23[year23['총_승객수'] == 0.00]
is23 = year23[year23['총_승객수'] !=0.00]

no24 = year24[year24['총_승객수'] == 0.00]
is24 = year24[year24['총_승객수'] !=0.00]


# 각 연도별로 0.0인 행정동과 아닌 행정동을 출력
print(f"\n2022년도 - 총_승객수가 0.0인 행정동:")
print(no22['행정동'].unique())

print(f"\n2022년도 - 총_승객수가 0.0이 아닌 행정동:")
print(is22['행정동'].unique())

print(f"\n2023년도 - 총_승객수가 0.0인 행정동:")
print(no23['행정동'].unique())

print(f"\n2023년도 - 총_승객수가 0.0이 아닌 행정동:")
print(is23['행정동'].unique())

print(f"\n2024년도 - 총_승객수가 0.0인 행정동:")
print(no24['행정동'].unique())

print(f"\n2024년도 - 총_승객수가 0.0이 아닌 행정동:")
print(is24['행정동'].unique())

is22
no22

print(no22.isna().sum())
print(no23.isna().sum())
print(no24.isna().sum())

print(is22.isna().sum())
print(is23.isna().sum())
print(is24.isna().sum())

# 년도별 데이터에 행정동 갯수,목록 확인

is22['행정동'].unique()
is22['행정동'].nunique()

is23['행정동'].unique()
is23['행정동'].nunique()

is24['행정동'].unique()
is24['행정동'].nunique()

# 22 : 219
# 23 : 220
# 24 : 221

print(is24.dtypes)

is22 
no22

is22.groupby(['연도', '분기', '행정동', '시군구'])['총_승객수'].mean()
is23.groupby(['연도', '분기', '행정동', '시군구'])['총_승객수'].mean()
is24.groupby(['연도', '분기', '행정동', '시군구'])['총_승객수'].mean()

# is_years = [is22, is23, is24]
# results = [df.groupby(['연도','분기','행정동'])['총_승객수'].mean().reset_index() for df in is_years]

# 데이터프레임 합치기 
year22_all = pd.concat([is22, no22],ignore_index=True)
year23_all = pd.concat([is23, no23],ignore_index=True)
year24_all = pd.concat([is24, no24],ignore_index=True)

print(year22_all)
print(year23_all)
print(year24_all)

# year24 3,4분기 삭제 

year24_q34 = year24_all[~((year24_all['연도'] == 2024) & (year24_all['분기'].isin([3,4])))]
print(year24_q34)

# print(year22_all.columns)
# print(year23_all.columns)
# print(year24_q34.columns)

# 각 데이터프레임의 병합 기준 컬럼에 NaN 값이 있는지 확인
# print(year22_all[['연도', '분기', '행정동', '시군구', '총_승객수']].isna().sum())
# print(year23_all[['연도', '분기', '행정동', '시군구', '총_승객수']].isna().sum())
# print(year24_q34[['연도', '분기', '행정동', '시군구', '총_승객수']].isna().sum())


year22_23 = pd.concat([year22_all, year23_all],ignore_index=True)
year_all = pd.concat([year22_23, year24_q34],ignore_index=True)

year_all

# 인덱스 삭제 , 연분기컬럼 생성
year_all['연분기'] = year_all['연도'].astype(str) + year_all['분기'].astype(str)
year_all

# 연도, 분기 컬럼 삭제 
year_all = year_all.drop(columns=['연도','분기'])

year_all
year_all.to_csv('year_all.csv', index=False, encoding='utf-8')
