==> JSW/8.zero_dong.ipynb <==
import numpy as np 
import pandas as pd 

year22 = pd.read_csv('y22_quarter_sum.csv')
year23 = pd.read_csv('y23_quarter_sum.csv')
year24 = pd.read_csv('y24_quarter_sum.csv')

year22
year23
year24

# 지하철_승객_수 컬럼명 바꾸기 

year_df = [year22, year23, year24]

for df in year_df: 
    df.rename(columns={'지하철_승객_수':'총_승객수'}, inplace=True)

for i, df in enumerate(year_df, start=22):
    print(f"year{i} 데이터프레임의 컬럼명:")
    print(df.head())

year22
year23
year24

# 총_승객수가 0.0인것과 값이 있는 행정동 목록을 각 년도로 나눠서 확인
year_df = [year22, year23, year24]
years = [2022, 2023, 2024]

for df, year in zip(year_df, years):
    zero_human = df[df['총_승객수'] == 0.0] # '총_승객수'가 0.0인 행정동을 필터링 

    is_human = df[df['총_승객수'] != 0.0] # '총_승객수'가 0.0이 아닌 행정동을 필터링 

    zero_dong = zero_human['행정동'].unique() # '총_승객수'가 0.0인 행정동 목록 
    print(df.isna().sum())

    is_dong = is_human['행정동'].unique() # '총_승객수'가 0.0이 아닌 행정동 목록 

    # 출력 

    # print(f"\n{year}년도")
    # print(f"총_승객수가 0.0인 행정동 목록:")
    # print(zero_dong)

    # print(f"\n{year}년도")
    # print(f"총_승객수가 0.0이 아닌 행정동 목록:")
    # print(is_dong)

중간에 None 값이 생겨서 코드 수정

# # 행정동, 값이 0.0인것과 값이 있는것을 각각 년도별 데이터프레임 csv로 저장 
# year_df = [year22, year23, year24]
# years = [2022, 2023, 2024]

# for df, year in zip(year_df, years):
#     zero_human = df[df['총_승객수'] == 0.0]
#     zero_dong = zero_human['행정동'].unique() # '총_승객수'가 0.0인 행정동 목록 
    

#     is_human = df[df['총_승객수'] != 0.0]
#     is_dong = is_human['행정동'].unique() # '총_승객수'가 0.0이 아닌 행정동 목록 

#     # 데이터프레임 생성 
#     zero_dong_df = pd.DataFrame({'연도': [year] * len(zero_dong), '행정동': zero_dong, '총_승객수':[0.0] * len(zero_dong)})
#     is_dong_df = pd.DataFrame({'연도': [year] * len(is_dong), '행정동':is_dong, '총_승객수':[0.0] * len(is_dong)})

#     # '총 승객수'가 0.0이 아닌 행정동 목록의 실제 '총_승객수' 값 사용 
#     zero_dong_sum = zero_human[['연도','분기','행정동','시군구','총_승객수']]
#     is_dong_sum = is_human[['연도', '분기','행정동','시군구','총_승객수']]

#     # csv 파일로 저장 
#     zero_dong_sum.to_csv(f'year_{year}_zero_dong.csv', index=False)
#     is_dong_sum.to_csv(f'year_{year}_is_dong.csv', index=False)

#     # 출력 (작업 완료 메시지)
#     print(f"\n{year}년도:")
#     print(f"총_승객수가 0.0인 행정동 목록 저장: 'year_{year}_zero_dong.csv'")
#     print(f"총_승객수가 0.0이 아닌 행정동 목록 저장: 'year_{year}_is_dong.csv'")

# 분기 변환

# 총_승객수가 0.0인것과 값이 있는것을 각각 년도별 데이터프레임 csv로 저장 
year_df = [year22, year23, year24]  # 연도별 데이터프레임 리스트
years = [2022, 2023, 2024]          # 해당 연도 리스트

# 분기 변환을 위한 매핑 딕셔너리
quarter_mapping = {'1-3월':'1분기',
                   '4-6월':'2분기',
                   '7-9월':'3분기',
                   '10-12월':'4분기',
                   '10-11월':'4분기'}

# 연도별 처리
for df, year in zip(year_df, years):
    # 총_승객수가 0.0인 행정동 목록
    zero_human = df[df['총_승객수'] == 0.0]
    zero_human['분기'] = zero_human['분기'].map(quarter_mapping)  # 분기 변환

    # 0.0인 행정동의 데이터
    zero_dong_sum = zero_human[['연도', '분기', '행정동','시군구', '총_승객수']]

    print(zero_dong_sum.isna().sum())

    # 총_승객수가 0.0이 아닌 행정동 목록
    is_human = df[df['총_승객수'] != 0.0]
    is_human['분기'] = is_human['분기'].map(quarter_mapping)  # 분기 변환

    # 0.0이 아닌 행정동의 데이터
    is_dong_sum = is_human[['연도', '분기', '행정동','시군구', '총_승객수',]]

    # CSV 파일로 저장
    zero_dong_sum.to_csv(f'year_{year}_zero_dong.csv', index=False)
    is_dong_sum.to_csv(f'year_{year}_is_dong.csv', index=False)

    # 출력 (작업 완료 메시지)
    print(f"\n{year}년도:")
    print(f"총_승객수가 0.0인 행정동 목록 저장: 'year_{year}_zero_dong.csv'")
    print(f"총_승객수가 0.0이 아닌 행정동 목록 저장: 'year_{year}_is_dong.csv'")


# # 분기 변환

# # 총_승객수가 0.0인것과 값이 있는것을 각각 년도별 데이터프레임 csv로 저장 
# year_df = [year22, year23, year24]  # 연도별 데이터프레임 리스트
# years = [2022, 2023, 2024]          # 해당 연도 리스트

# # 분기 변환을 위한 매핑 딕셔너리
# quarter_mapping = {'1-3월':'1분기',
#                    '4-6월':'2분기',
#                    '7-9월':'3분기',
#                    '10-12월':'4분기'}

# # 연도별 처리
# for df, year in zip(year_df, years):
#     # 총_승객수가 0.0인 행정동 목록
#     zero_human = df[df['총_승객수'] == 0.0]
#     zero_human.loc[:,'분기'] = zero_human['분기'].map(quarter_mapping)  # 분기 변환

#     # 0.0인 행정동의 데이터
#     zero_dong_sum = zero_human[['연도', '분기', '행정동','시군구', '총_승객수']]

#     # 총_승객수가 0.0이 아닌 행정동 목록
#     is_human = df[df['총_승객수'] != 0.0]
#     is_human.loc[:,'분기'] = is_human['분기'].map(quarter_mapping)  # 분기 변환

#     # 0.0이 아닌 행정동의 데이터
#     is_dong_sum = is_human[['연도', '분기', '행정동','시군구', '총_승객수',]]

#     # CSV 파일로 저장
#     zero_dong_sum.to_csv(f'year_{year}_zero_dong.csv', index=False)
#     is_dong_sum.to_csv(f'year_{year}_is_dong.csv', index=False)

#     # 출력 (작업 완료 메시지)
#     print(f"\n{year}년도:")
#     print(f"총_승객수가 0.0인 행정동 목록 저장: 'year_{year}_zero_dong.csv'")
#     print(f"총_승객수가 0.0이 아닌 행정동 목록 저장: 'year_{year}_is_dong.csv'")


zero24 = pd.read_csv('year_2024_zero_dong.csv')

print(zero24.isna().sum())

nan = zero24[zero24['분기'].isna()]
